+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.3
+
+* Add prev / next buttons let user switch medias.
+* Add class JkVideoPlaylistPlayer to play a playlist in a simple way.
+* Support play 2 more videos at the same time.
+* many bug fix
+
 ## 1.0.2
 
 * [Android / iOS] force lock orientation by video aspect ratio in fullscreen mode.